Fix 30 MSVC warnings unmasked once the Qt header failure was cleared
With C4702 in qnumeric.h no longer aborting the Windows build, MSVC reached
the rest of the tree and reported 30 warnings in our own code.

C4805 (unsafe bool/int mix), 15 sites in BrainView::clearProbeVisualization:
QMap::remove returns the number of entries removed, not a bool, and each
result was OR-ed into a bool with |=. Sum the counts into a qsizetype and
compare against zero instead.

C4458 (declaration hides class member), 7 sites: parameters and locals named
'data', which hides QWidget::data in any widget derived class. Affected
BrainView::pushRealtimeSourceData/pushRealtimeSensorData,
DataWindow::loadFiffBuffer/onBlockLoaded/appendVirtualChannels/
applyUserDefinedFilter and a local in MainWindow::computeIca. Renamed with
matching header updates. Note clang cannot see this class of shadowing:
neither -Wshadow nor -Wshadow-field reports a local hiding a private base
class member, which is why a tree wide clang sweep reported zero sites.

C4267 (size_t to smaller type), 6 sites: EventModel row counts, BrainSurface
buffer sizes, DipoleObject index count and a BrainRenderer buffer size, all
now explicitly cast.

C4305 (double truncated to float), 2 sites: QColor::setAlphaF takes a float.

Verified: clang build 0 warnings, 206/206 ctest pass.

+++ src/libraries/disp3D/renderable/brainsurface.cpp
@@ -523,8 +523,8 @@ void BrainSurface::updateBuffers(QRhi *rhi, QRhiResourceUpdateBatch *u)
     if (!m_gpu->dirty && !needsCreate) return;
 #endif
 
-    const quint32 vbufSize = m_vertexData.size() * sizeof(VertexData);
-    const quint32 ibufSize = m_indexData.size() * sizeof(uint32_t);
+    const quint32 vbufSize = static_cast<quint32>(m_vertexData.size() * sizeof(VertexData));
+    const quint32 ibufSize = static_cast<quint32>(m_indexData.size() * sizeof(uint32_t));
 
 #ifdef __EMSCRIPTEN__
     if (!m_gpu->vertexBuffer) {
